Feat: 게시글 상세 조회 구현

+++ app/src/main/java/com/example/umc_closit/data/remote/post/PostResponse.kt
@@ -33,8 +33,8 @@ data class PostDeleteResponse(
 )
 
 data class ItemTag(
-    val x: Int,
-    val y: Int,
+    val x: Float,
+    val y: Float,
     val content: String
 )
